requirements.txt
asgiref==3.8.1
boto3==1.35.40
botocore==1.35.40
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==0.50.0
django-storages==1.14.4
djangorestframework==3.15.2
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
